+++ app/Console/Commands/Process/ProcessSpamEmail.php
@@ -3,7 +3,9 @@
 namespace App\Console\Commands\Process;
 
 use App\IronPort\IronPortSpamEmail;
+use Carbon\Carbon;
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Log;
 
 class ProcessSpamEmail extends Command
 {
@@ -38,11 +40,18 @@ class ProcessSpamEmail extends Command
      */
     public function handle()
     {
+        /*
+         * [2] Process spam emails into database
+         */
+
+        Log::info(PHP_EOL.'***********************************'.PHP_EOL.'* Starting spam email processing! *'.PHP_EOL.'***********************************');
+
         $contents = file_get_contents(storage_path('app/collections/spam.json'));
         $spam_emails = \Metaclassing\Utility::decodeJson($contents);
 
         foreach ($spam_emails as $spam) {
             $reasons = '';
+
             // cycle through reasons
             foreach ($spam['reason'] as $reason) {
                 // grab policy array and convert to a ';' separated string
@@ -75,43 +84,24 @@ class ProcessSpamEmail extends Command
             $quarantines = implode('; ', $spam['quarantine_names']);
             $recipients = implode('; ', $spam['recipients']);
 
-            // attempt to find existing record and update
-            $updated = IronPortSpamEmail::where('mid', $spam['mid'])->update([
-                'mid'                 => $spam['mid'],
-                'subject'             => $spam['subject'],
-                'size'                => $spam['size'],
-                'quarantine_names'    => $quarantines,
-                'time_added'          => $datetime,
-                'reason'              => $reasons,
-                'recipients'          => $recipients,
-                'sender'              => $spam['sender'],
-                'esa_id'              => $spam['esa_id'],
-                'data'                => \Metaclassing\Utility::encodeJson($spam),
-            ]);
+            $spam_model = IronPortSpamEmail::updateOrCreate(
+                ['mid'                 => $spam['mid']],
+                ['subject'             => $spam['subject']],
+                ['size'                => $spam['size']],
+                ['quarantine_names'    => $quarantines],
+                ['time_added'          => $datetime],
+                ['reason'              => $reasons],
+                ['recipients'          => $recipients],
+                ['sender'              => $spam['sender']],
+                ['esa_id'              => $spam['esa_id']],
+                ['data'                => \Metaclassing\Utility::encodeJson($spam)]
+            );
 
-            // if not updated then create new
-            if (!$updated) {
-                echo 'creating spam record for: '.$spam['sender'].PHP_EOL;
-                $new_spam = new IronPortSpamEmail();
-
-                $new_spam->mid = $spam['mid'];
-                $new_spam->subject = $spam['subject'];
-                $new_spam->size = $spam['size'];
-                $new_spam->quarantine_names = $quarantines;
-                $new_spam->time_added = $datetime;
-                $new_spam->reason = $reasons;
-                $new_spam->recipients = $recipients;
-                $new_spam->sender = $spam['sender'];
-                $new_spam->esa_id = $spam['esa_id'];
-                $new_spam->data = \Metaclassing\Utility::encodeJson($spam);
-
-                $new_spam->save();
-            } else {
-                echo 'updated spam record for: '.$spam['sender'].PHP_EOL;
-            }
+            // touch spam model to update the "updated_at" timestamp in case nothing was changed
+            $spam_model->touch();
         }
 
-        //$this->processDeletes();
+        $this->processDeletes();
     }
 
     /**
@@ -121,15 +111,19 @@ class ProcessSpamEmail extends Command
      */
     public function processDeletes()
     {
-        $today = new \DateTime('now');
-        $pastdate = $today->modify('-3 month');
-        $delete_date = $pastdate->format('Y-m-d H:i:s');
+        $delete_date = Carbon::now()->subMonths(6);
+        Log::info('spam delete date: '.$delete_date);
 
-        $spam_emails = IronPortSpamEmail::where('updated_at', '<=', $delete_date)->get();
+        $spam_emails = IronPortSpamEmail::all();
 
         foreach ($spam_emails as $spam) {
-            echo 'deleting spam record: '.$spam->id.PHP_EOL;
-            $spam->delete();
+            $updated_at = Carbon::createFromFormat('Y-m-d H:i:s', $spam->updated_at);
+
+            if ($updated_at->lt($delete_date))
+            {
+                Log::info('deleting spam record: '.$spam->id);
+                $spam->delete();
+            }
         }
     }
 }    // end of ProcessSpamEmail command class
